Add files via upload
Hey guys this is a theoretical Calculator that I made. It should work. I used components from the internet and had to change some stuff up but the idea is the same. this was used to help me study calculator design on what works. Will attempt at my own design next.